requirements.txt
asgiref==3.4.1
automium==0.2.6
automium-web==0.1.1
click==8.0.3
codegen==1.0
colorama==0.4.4
Django==3.2.9
django-crispy-forms==1.13.0
django-email-users==1.1.0
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
legacy==0.1.6
MarkupSafe==2.0.1
patterns==0.3
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
views==0.3
Werkzeug==2.0.2
whitenoise==6.1.0
